Minor Fixes & event OpenGraph
