+++ unlock-app/app/frames/event/Components/Description.tsx
@@ -2,6 +2,8 @@ import React from 'react'
 import Container from './Container'
 import { FrameButton, FrameImage } from 'frames.js/next/server'
 import { config } from '../../../../src/config/app'
+import removeMd from 'remove-markdown'
+import { truncateString } from '~/utils/truncateString'
 
 export const Description = ({
   event,
@@ -12,23 +14,6 @@ export const Description = ({
   previousFrame: any
   state: any
 }) => {
-  const removeMd = require('remove-markdown')
-
-  function truncateString(text: string, maxLength: number): string {
-    if (text.length <= maxLength) {
-      return text
-    }
-
-    const shortenedString = text.substring(0, maxLength)
-    const lastSpaceIndex = shortenedString.lastIndexOf(' ')
-
-    if (lastSpaceIndex !== -1) {
-      return shortenedString.substring(0, lastSpaceIndex) + '...'
-    } else {
-      return shortenedString + '...'
-    }
-  }
-
   return (
     <Container slug={event.slug} previousFrame={previousFrame} state={state}>
       <FrameImage>
